+++ drt/sources/redshift.py
@@ -22,8 +22,6 @@ import os
 from collections.abc import Iterator
 from typing import Any
 
-from psycopg2 import sql
-
 from drt.config.credentials import ProfileConfig, RedshiftProfile
 
 
@@ -45,7 +43,7 @@ class RedshiftSource:
             cur = conn.cursor()
             # Set search_path to the configured schema
             if config.schema:
-                cur.execute(sql.SQL("SET search_path TO {}").format(sql.Identifier(config.schema)))
+                cur.execute("SET search_path TO %s", (config.schema,))
             cur.execute(query)
             columns = [desc[0] for desc in cur.description]
             for row in cur.fetchall():