+++ integrations/dagster-drt/dagster_drt/__init__.py
@@ -1,3 +1,4 @@
-from dagster_drt.assets import drt_assets
+from dagster_drt.assets import DrtConfig, drt_assets
+from dagster_drt.translator import DagsterDrtTranslator
 
-__all__ = ["drt_assets"]
+__all__ = ["DagsterDrtTranslator", "DrtConfig", "drt_assets"]